[#1796]Move DatasetImporter & CoerceToSql to a.l.protocols
