+++ backend/src/akvo/lumen/lib/import/common.clj
@@ -2,57 +2,9 @@
   (:require [clojure.java.io :as io]
             [clojure.java.jdbc :as jdbc]
             [clojure.string :as str]
+            [akvo.lumen.protocols :as p]
             [org.akvo.resumed :as resumed])
   (:import [org.postgis PGgeometry]))
-
-(defprotocol DatasetImporter
-  "
-  A protocol for importing datasets into Lumen. A typical implementation
-  should also implement `java.io.Closeable` since some data sources are
-  backed by resources that need to be released.
-
-  Example:
-  (reify
-    Closeable
-    (close [this])
-
-    DatasetImporter
-    (columns [this]
-      [{:id :a :type :text :title \"A\"}
-       {:id :b :type :number :title \"B\"}
-       {:id :c :type :date :title \"C\"}])
-    (records [this]
-      [{:a \"foo\"
-        :b 42
-        :c (Instant/now)}
-       {:a \"bar\"
-        :b 3.14
-        :c (Instant/now)}
-  "
-
-  (columns [this]
-    "Returns a sequence of column specifications of the dataset to be imported.
-     A column specification is a map with keys
-
-     Required:
-       :type - The lumen type of the column. Currently :text, :number, :date, :geoshape or :geopoint
-       :title - The title of the column
-       :id - The internal id of the column (as a keyword). The id must be
-             lowercase alphanumeric ([a-z][a-z0-9]*)
-
-     Optional:
-       :key - True if this column is required to be non-null and unique")
-  (records [this]
-    "Returns a sequence of record data. A record is a map of column ids to values.
-     The type of the value depends on the type of the column where
-
-       :text - java.lang.String
-       :number - java.lang.Number
-       :date - java.time.Instant
-       :geoshape - Geoshape
-                   Well-known text (WKT) shape (POLYGON or MULTIPOLYGON)
-       :geopoint - Geopoint
-                   Well-known text (WKT) shape (POINT)"))
 
 (defn dispatch-on-kind [spec]
   (let [kind (get spec "kind")]
@@ -123,10 +75,7 @@
 
 (defrecord Geopoint [wkt-string])
 
-(defprotocol CoerceToSql
-  (coerce [this]))
-
-(extend-protocol CoerceToSql
+(extend-protocol p/CoerceToSql
   java.lang.String
   (coerce [value] value)
   java.lang.Number
@@ -148,7 +97,7 @@
 (defn coerce-to-sql [record]
   (reduce-kv
    (fn [result k v]
-     (assoc result k (when v (coerce v))))
+     (assoc result k (when v (p/coerce v))))
    {}
    record))
