+++ backend/src/akvo/lumen/lib/import/flow_common.clj
@@ -1,6 +1,5 @@
 (ns akvo.lumen.lib.import.flow-common
   (:require [akvo.commons.psql-util :as pg]
-            [akvo.lumen.lib.import.common :as import]
             [cheshire.core :as json]
             [clj-http.client :as http]
             [clojure.java.jdbc :as jdbc]
