+++ backend/src/akvo/lumen/lib/import/flow_v2.clj
@@ -1,6 +1,5 @@
 (ns akvo.lumen.lib.import.flow-v2
   (:require [akvo.commons.psql-util :as pg]
-            [akvo.lumen.lib.import.common :as import]
             [akvo.lumen.lib.import.flow-common :as flow-common]
             [cheshire.core :as json]
             [clj-http.client :as http]
